+++ upload_crop.php
@@ -229,8 +229,12 @@ if(isset($_POST["upload"])) {
 
 		}
 
-		header("location:".$_SERVER["PHP_SELF"]);
-		exit();
+		header("location:upload_crop.php?c=confirm");
+		
+		
+		
+		
+		//exit();
 	}
 }
 
@@ -246,7 +250,10 @@ if (isset($_POST["upload_thumbnail"]) && strlen($large_photo_exists)>0) {
 	$scale = $thumb_width/$w;
 	$cropped = resizeThumbnailImage($thumb_image_location, $large_image_location,$w,$h,$x1,$y1,$scale);
 	//Reload the page again to view the thumbnail
-	header("location:".$_SERVER["PHP_SELF"]);
+	
+	
+	session_destroy();
+	header("location: upload_crop.php?c=confirm");
 	exit();
 }
 
@@ -265,6 +272,38 @@ if($_GET['a']=="delete" && strlen($_GET['t'])>0){
 }
 ?>
 <main>
+
+<?php
+
+// DISPLAY ERROR MESSAGE IF THE USER WANTS TO DELETE THE CURRENT IMAGE
+$error2 = $_REQUEST['error2'];
+
+if (	$error2=='1'	) {
+
+	echo "<section class='info'>
+
+		<p>Error: The image you want to delete is your current profile pictire.</p>
+
+	</section>";
+
+}
+
+
+// CONFIRM MESSAGE
+$c = $_REQUEST['c'];
+
+if ($c=='confirm') {
+	
+	echo "<section class='info'>
+
+		<p>Your profile picture has been updated.</p>
+
+	</section>";
+	
+}
+
+?>
+
   <section>
     <h1>Post New Profile Picture</h1>
     <?php
@@ -367,14 +406,44 @@ if($_GET['a']=="delete" && strlen($_GET['t'])>0){
 
 	$deleteImgOwnersID = $deleteQueryArray['owners_id'];
 	$deleteImgLoc = $deleteQueryArray['location'];
+	
+	
 		// SECURITY CHECK IF THE USER IS THE OWNER IF THE IMAGE
 		if($deleteImgOwnersID == $usersIDCookie) {
+			
+			$imageCheckQuery = mysqli_query($con, "SELECT profile_picture FROM users WHERE id = '$usersIDCookie' ");
+			$imageCheckQueryResult = mysqli_fetch_array($imageCheckQuery);
+			$currentBannerIMG = $imageCheckQueryResult['profile_picture'];
+			
+			// CHECK IF THE IMAGE IS NOT THE CURRENT SET IMAGE
+			if ($deleteImgLoc!==$currentBannerIMG) {
+			
+			
+			
 				// DELETE THE ROW FROM THE DATABASE
 				mysqli_query($con, "DELETE FROM used_profile_pics WHERE id='$delete'; ");
-				// DELETE THE IMAGE FROM THE IMAGES FOLDER
-				$imgPath = "$deleteImgLoc";
-				unlink($imgPath);
+				
+				
+				// CHECK IF THE IMAGE IS THE DEAFULT IMAGE
+				if ($deleteImgLoc!=='images/profile-pictures/default_profile.jpg') {
+				
+					// DELETE THE IMAGE FROM THE IMAGES FOLDER
+					$imgPath = "$deleteImgLoc";
+					unlink($imgPath);
+				
+				}
+				
 				header("location: upload_crop.php"); //REFRESH PAGE
+				
+			} else {
+				
+				// DISPLAY ERROR MESSAGE
+				header("location: upload_crop.php?error2=1");
+				
+				
+			}
+				
+				
 			} else { // NOT THE USERS IMAGE
 				header("location: upload_crop.php");
 			}
@@ -430,18 +499,8 @@ if($_GET['a']=="delete" && strlen($_GET['t'])>0){
 			</div>
 		</div>";
     }
-    if(!empty($_REQUEST['c'])) {
-      $confirm = "Your profile picture has been updated";
-    }
+    
     ?>
   </section>
-  <?php
-  if($confirm != "") {
-    echo "
-    <section class='info'>
-      <p>$confirm</p>
-    </section>";
-  }
-  ?>
 </main>
 <?php include_once($_SERVER["DOCUMENT_ROOT"]."/Pompay/footer.php"); ?>
